1. Kullanici giris ve çıkış işlemleri gerçekleştirildi. Tüm testleri yapıldı ve başarı ile geçmiştir. 2. Kullanıcı kayıt işlemleri gerçekleştirildi.

+++ MobilyaETicaret.Service/Services/FotograflarService.cs
@@ -23,7 +23,7 @@ namespace MobilyaETicaret.Service.Services
 			_mapper = mapper;
 		}
 
-		public async Task<int> FotografEkleAsync(string fotografYolu, string adi, string fotografAciklamasi, byte fotografSirasi, int urunId, bool aktifMi, DateTime eklemeTarihi, DateTime guncellemeTarihi)
+		public async Task<int> FotografEkleAsync(string fotografYolu, string adi, string fotografAciklamasi, int fotografSirasi, int urunId, bool aktifMi, DateTime eklemeTarihi, DateTime guncellemeTarihi)
 		{
 			try
 			{
